+++ src/transaction/transaction.go
@@ -4,9 +4,9 @@ import "global"
 
 //事务对象
 type Trx struct {
-	id             int64
-	startTime      int32
-	endTime        int32
+	id             int
+	startTime      int
+	endTime        int
 	isolationLevel string
 	firstReadView  ReadView //第一个视图
 }
